pyproject.toml
[project]
name = "ai-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "openai>=2.24.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
